src/appli12/pyproject.toml
[project]
name = "application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.4",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.8.0",
]
